requirements.txt
numpy==2.0.0
opencv_python==4.10.0.84
Pillow==10.3.0
